Rename: svg 파일 폴더화
