+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import Header from './components/Header';
-import CakeWay from '../../public/cake-way.svg';
-import Alarm from '../../public/alarm.svg';
+import CakeWay from '../../public/Header/cake-way.svg';
+import Alarm from '../../public/Header/alarm.svg';
 import Image from 'next/image';
 import InputField from './components/InputField';
 import CakePick from './components/home/CakePick';
